Move enrichment and references under the `type` package
