+++ base/src/main/java/io/spine/type/enrichment/BuiltIn.java
@@ -18,12 +18,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.code.proto.enrichment;
+package io.spine.type.enrichment;
 
 import com.google.errorprone.annotations.Immutable;
 import com.google.protobuf.Descriptors.Descriptor;
 import io.spine.base.MessageContext;
-import io.spine.code.proto.ref.TypeRef;
+import io.spine.type.ref.TypeRef;
 
 import java.util.Optional;
