+++ base/src/main/java/io/spine/type/enrichment/EnrichmentType.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.code.proto.enrichment;
+package io.spine.type.enrichment;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
@@ -27,9 +27,9 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Message;
-import io.spine.code.proto.ref.TypeRef;
 import io.spine.type.KnownTypes;
 import io.spine.type.MessageType;
+import io.spine.type.ref.TypeRef;
 
 import java.util.Collection;
 import java.util.List;
