+++ base/src/main/java/io/spine/type/enrichment/FieldDef.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.code.proto.enrichment;
+package io.spine.type.enrichment;
 
 import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.Immutable;
