+++ base/src/main/java/io/spine/type/ref/DirectTypeRef.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.code.proto.ref;
+package io.spine.type.ref;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
