+++ base/src/test/java/io/spine/type/KnownTypesTest.java
@@ -32,12 +32,12 @@ import com.google.protobuf.StringValue;
 import com.google.protobuf.Timestamp;
 import io.spine.base.Error;
 import io.spine.code.java.ClassName;
-import io.spine.code.proto.ref.TypeRef;
 import io.spine.option.EntityOption;
 import io.spine.option.IfMissingOption;
 import io.spine.test.types.KnownTask;
 import io.spine.test.types.KnownTaskId;
 import io.spine.test.types.KnownTaskName;
+import io.spine.type.ref.TypeRef;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
