+++ base/src/test/java/io/spine/type/enrichment/EnrichmentTypeTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.code.proto.enrichment;
+package io.spine.type.enrichment;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.truth.IterableSubject;
